Rename Playthrough to Search, pt. 1.
