+++ Main.py
@@ -29,7 +29,7 @@ from N64Patch import create_patch_file, apply_patch_file
 from SettingsList import setting_infos, logic_tricks
 from Rules import set_rules, set_shop_rules
 from Plandomizer import Distribution
-from Playthrough import Playthrough, RewindablePlaythrough
+from Search import Playthrough, RewindablePlaythrough
 from EntranceShuffle import set_entrances
 from LocationList import set_drop_location_names
